device_management/requirements.txt
flask>=2.0.0
psycopg2-binary>=2.9.0
flask_jwt_extended>=4.0.0
